+++ internal/pkg/warewulfd/server/server.go
@@ -64,6 +64,7 @@ func configureRootHandler(apiHandler http.Handler) *slashFix {
 	wwHandler.HandleFunc("/system/", warewulfd.HandleSystemOverlay)
 	wwHandler.HandleFunc("/runtime/", warewulfd.HandleRuntimeOverlay)
 	wwHandler.HandleFunc("/status", warewulfd.HandleStatus)
+	wwHandler.HandleFunc("/files/", warewulfd.HandleFiles)
 
 	/* Deprecated */
 	wwHandler.HandleFunc("/container/", warewulfd.HandleImage)